Initial commit
We want a new shiny Telegram bot. Let's create it!

As promised previously, this is the initial git tree for the new
GNU/Weeb Telegram bot project. I encourage everyone to contribute to
this project. I am looking forward to your contribution.

I call it GNUWeebBot2. It's the second attempt to create a Telegram bot
project for the GNU/Weeb community; the former GNUWeebBot doesn't
perform really well. This one will be the better version of it.

Currently, this project only runs on Linux.

Link: https://t.me/GNUWeeb/695804
Acked-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>