core/thread: Don't unlock-relock in cond_wait()
Avoid unlock-relock in cond_wait() by introducing a new field, a pointer
to std::unique_lock<std::mutex>, in the mutex struct. Construct it with
std::adopt_lock, and check whether this pointer is NULL when unlocking.
If not, unlock it via that unique lock, delete it, and set the pointer
back to NULL.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>